Обновлены модули шаблонов игры
