--- js/templates/header/header-screen.js
@@ -1,12 +1,0 @@
-import HeaderView from './header-view';
-import {resetGame} from '../../utils/util';
-
-const renderHeader = (game) => {
-  const header = new HeaderView(game);
-  header.onButtonClick = () => {
-    resetGame();
-  };
-  return header;
-};
-
-export default renderHeader;
