+++ js/templates/header/header-view.js
@@ -1,4 +1,5 @@
 import AbstractView from '../../utils/abstract-view';
+import Application from '../../application';
 
 class HeaderView extends AbstractView {
   constructor(game) {
@@ -24,12 +25,8 @@ class HeaderView extends AbstractView {
 
   bind() {
     this.element.querySelector(`.back`).addEventListener(`click`, () => {
-      this.onButtonClick();
+      Application.showGreeting();
     });
-  }
-
-  onButtonClick() {
-
   }
 }
