+++ js/templates/intro/intro-view.js
@@ -1,4 +1,6 @@
 import AbstractView from '../../utils/abstract-view';
+import FooterView from '../footer/footer-view';
+import Application from '../../application';
 
 class IntroView extends AbstractView {
   constructor() {
@@ -11,17 +13,14 @@ class IntroView extends AbstractView {
           <h1 class="intro__asterisk">*</h1>
           <p class="intro__motto"><sup>*</sup> Это не фото. Это рисунок маслом нидерландского художника-фотореалиста Tjalf Sparnaay.</p>
         </div>
-      </div>`;
+      </div>
+      ${new FooterView().template}`;
   }
 
   bind() {
     this.element.querySelector(`.intro__asterisk`).addEventListener(`click`, () => {
-      this.onButtonClick();
+      Application.showGreeting();
     });
-  }
-
-  onButtonClick() {
-
   }
 }
