--- js/templates/rules/rules-screen.js
@@ -1,13 +1,0 @@
-import {changeView} from '../../utils/util';
-import nextScreen from '../game/game-screen';
-import RulesView from './rules-view';
-
-const renderRulesScreen = () => {
-  const screen = new RulesView();
-  screen.onFormSubmit = () => {
-    changeView(nextScreen);
-  };
-  return screen;
-};
-
-export default renderRulesScreen;
