--- js/templates/statistics/statistic-view.js
@@ -1,32 +1,0 @@
-import AbstractView from '../../utils/abstract-view.js';
-import StatsView from './stats-view.js';
-import getElementFromTemplate from '../../utils/create-elem';
-import renderHeader from '../header/header-screen.js';
-import {BonusPoint} from '../../data/count-points.js';
-
-class StatisticsView extends AbstractView {
-  constructor(results) {
-    super();
-    this.results = results;
-    this.block1 = new StatsView(this.results[0], BonusPoint, 1);
-    this.block2 = new StatsView(this.results[1], BonusPoint, 2);
-    this.block3 = new StatsView(this.results[2], BonusPoint, 3);
-  }
-
-  get template() {
-    return `<div class="result">
-      <h1>${this.results[0].victory ? `Победа!` : `Поражение :(`}</h1>
-      ${this.block1.template}
-      ${this.block2.template}
-      ${this.block3.template}
-      </div>`;
-  }
-
-  render() {
-    const element = getElementFromTemplate(this.template);
-    element.insertAdjacentElement(`afterbegin`, renderHeader().element);
-    return element;
-  }
-}
-
-export default StatisticsView;
