--- js/templates/statistics/statistics-screen.js
@@ -1,6 +1,0 @@
-import {totalResults} from '../../data/data.js';
-import StatisticsView from './statistic-view.js';
-
-const statsScreen = new StatisticsView(totalResults).element;
-
-export default statsScreen;
